added editor top section
